+++ CHANGELOG.md
@@ -1,9 +1,14 @@
+Stage
+==========
+
+### Features
+* Adiciona estrutura para inserção, listagem, edição e deleção lógica de cargos
+
 Change Log
 ==========
 
 <a name="v0.0.97"></a>
 ## Version [v0.0.97](https://github.com/luizfp/PrologWebService/compare/v0.0.96...v0.0.97) (release-date) [unreleased]
-
 
 <a name="v0.0.96"></a>
 ## Version [v0.0.96](https://github.com/luizfp/PrologWebService/compare/v0.0.95...v0.0.96) (2019-06-21)